+++ reporting-pipeline-service/src/main/java/gov/cdc/nbs/report/pipeline/observation/service/ObservationService.java
@@ -8,7 +8,7 @@ import static gov.cdc.etldatapipeline.commonutil.UtilHelper.extractValue;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import gov.cdc.etldatapipeline.commonutil.DataProcessingException;
 import gov.cdc.etldatapipeline.commonutil.NoDataException;
-import gov.cdc.nbs.report.pipeline.observation.service.observation.ObservationProcessor;
+
 import java.util.NoSuchElementException;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutorService;
