+++ reporting-pipeline-service/src/test/java/gov/cdc/nbs/report/pipeline/observation/repository/NrtObservationWriterTest.java
@@ -1,4 +1,4 @@
-package gov.cdc.nbs.report.pipeline.observation.service.observation;
+package gov.cdc.nbs.report.pipeline.observation.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -13,6 +13,7 @@ import gov.cdc.nbs.report.pipeline.observation.model.dto.observation.Observation
 import gov.cdc.nbs.report.pipeline.observation.model.dto.observation.ObservationNumeric;
 import gov.cdc.nbs.report.pipeline.observation.model.dto.observation.ObservationReason;
 import gov.cdc.nbs.report.pipeline.observation.model.dto.observation.ObservationTxt;
+
 import java.text.SimpleDateFormat;
 import java.util.List;
 import java.util.Map;
